Fixed crash when selecting an employee index (#122)
xxxx

Issue: While in the employee information screen, we can still select an
employee when inputting their index in the menu. This makes us to go into
an invalid state and crash.

Fix: Added checking to make sure we only interpret the input as an employee
index when we are in the employee management screen.